Fixed bugs, added feature
There were 2 bugs in the algorithm in addition the differential was changed to compute based on time deltas rather than assuming equal time intervals.  One of the bugs was in determining the magnitude of other slopes relative to the size of the minimum. And the second was a quirk of python where the list was appending with references to the same stats dict instead of copying a new set of data each time.