Update with full filename
There was a bug that uploaded the cvs files without the 6 digits preceding the .csv in the name. Rather than upload a new set with a slightly new name, here are two files as they should be named after running convert.py.  These files are used in the hard coded example filenames.